requirements-test.txt
# you can find the project dependencies in setup.py
.

pytest==3.4.2
pytest-cov==2.5.1
pytest-sugar==0.9.1
pytest-django==3.1.2